% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppdummy}
\alias{cppdummy}
\title{Read a given file in sparse (LIBSVM) format to dense R matrix and R vector.}
\usage{
cppdummy(i)
}
\arguments{
\item{filename}{the filename of the data in sparse file format. The Tilde Character isn't supported for f filename!}

\item{verbose}{show verbose output?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\value{
the data is read into an R matrix and an R vector, containing the data
                                    and the labels. note, that these are not in sparse format, but are dense.
}
\description{
Read a given file in sparse (LIBSVM) format to dense R matrix and R vector.
}
\note{
this routine is nearly a 1:1 adoptation from the LIBSVM original code.
}
\keyword{IO}

