% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cascadesvm.R
\name{cascadesvm}
\alias{cascadesvm}
\title{cascade SVM}
\usage{
cascadesvm(X, Y, k, epochs = 1, train.param = NULL, verbose = FALSE, ...)
}
\arguments{
\item{X}{training input}

\item{Y}{labels}

\item{k}{number of initial subsets.}

\item{epochs}{number of epochs to run}
}
\value{
svm models, containing: -final model (libsvm model), -?
}
\description{
we do not restrict the number of subsets. in case at any point there is an uneven
number of splits we combine all of them normally (two by two), except for the last
three subsets. these we combine directly.
}
\details{
FIXME: -stratified sampling,
}

